+++ src/test/java/com/itextpdf/rups/view/itext/treenodes/asn1/correctors/pdf/RevocationInfoArchivalCorrectorTest.java
@@ -40,7 +40,7 @@
     For more information, please contact iText Software Corp. at this
     address: sales@itextpdf.com
  */
-package com.itextpdf.rups.view.itext.treenodes.asn1.correctors.x509.attributes;
+package com.itextpdf.rups.view.itext.treenodes.asn1.correctors.pdf;
 
 import com.itextpdf.rups.view.itext.treenodes.asn1.AbstractAsn1TreeNode;
 import com.itextpdf.rups.view.itext.treenodes.asn1.Asn1TestUtil;
